+++ src/StatusCollection/StatusCollection.hpp
@@ -8,7 +8,7 @@
 class StatusCollection
 {
 private:
-	size_t		_statusTypes;
+	int		_statusTypes;
 	std::string	_path;
 public:
 	StatusCollection();
@@ -16,7 +16,7 @@ public:
 	StatusCollection(const StatusCollection& other);
 	StatusCollection& operator=  (const StatusCollection& other);
 	~StatusCollection();
-	size_t& getStatusTypes();
+	int& getStatusTypes();
 	std::string &getPath();
 	StatusCollection& setStatusTypes(const size_t statusType);
 	StatusCollection& setPath(const std::string& path);
